v1/requirements.txt
# FinCat Requirements

# File monitoring
watchdog==3.0.0

# Excel handling
xlrd==2.0.1          # Read legacy .xls files
openpyxl==3.1.2      # Read/write modern .xlsx files

# AI integration
anthropic>=0.75.0    # Claude API (official SDK)

# Configuration
PyYAML==6.0.1        # config.yaml parsing
python-dotenv==1.0.0 # .env file loading

# Optional (recommended for better UX)
colorama==0.4.6      # Cross-platform colored output
tqdm==4.66.1         # Progress bars
